+++ ui/src/types/annotations.ts
@@ -1,0 +1,8 @@
+export interface AnnotationInterface {
+  id: string
+  startTime: number
+  endTime: number
+  text: string
+  type: string
+  links: {self: string}
+}