+++ ui/src/utils/formatting.ts
@@ -117,23 +117,6 @@ export const numberValueFormatter = (
   return `${prefix}${label}${suffix}`
 }
 
-export const formatBytes = (bytes: number) => {
-  if (bytes === 0) {
-    return '0 Bytes'
-  }
-
-  if (!bytes) {
-    return null
-  }
-
-  const k = 1000
-  const dm = 2
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`
-}
-
 export const formatRPDuration = (duration: string | null): string => {
   if (!duration) {
     return