+++ ui/webpack/vendor.config.js
@@ -2,8 +2,20 @@ const path = require('path')
 const webpack = require('webpack')
 const packages = require('../package.json')
 const dependencies = packages.dependencies
+const ProgressBarPlugin = require('progress-bar-webpack-plugin')
+
+const stats = {
+  colors: true,
+  children: false,
+  modules: false,
+  version: false,
+  warnings: false,
+  assetsSort: '!size',
+}
 
 module.exports = {
+  mode: 'development',
+  stats,
   node: {
     fs: 'empty',
     module: 'empty',
@@ -22,6 +34,7 @@ module.exports = {
     library: '[name]',
   },
   plugins: [
+    new ProgressBarPlugin(),
     new webpack.DllPlugin({
       name: '[name]',
       path: './build/[name].dll.json',